Backend/requirements.txt
# ============================================================
# Framify Backend — Python Dependencies
# ============================================================

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# Database
pymongo==4.8.0
motor==3.5.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.27.0

# AI Copy Generation
openai==1.40.0

# Export / ZIP
aiofiles==24.1.0

# CORS
python-multipart==0.0.9

# Validation
pydantic==2.8.0
pydantic-settings==2.4.0
